Split things into smaller files. Cleaner version of SettingsValue.
