+++ Sources/Builder/Configuration.swift
@@ -6,204 +6,9 @@
 
 import Foundation
 
-struct SettingsValue : Equatable, Decodable, ExpressibleByArrayLiteral, ExpressibleByStringLiteral {
-    enum Value : Equatable {
-        case string(String)
-        case list([String])
-    }
-    
-    static func == (lhs: SettingsValue, rhs: SettingsValue) -> Bool {
-        return lhs.value == rhs.value
-    }
-    
-    
-    typealias ArrayLiteralElement = String
-    typealias StringLiteralType = String
-    
-    let value: Value
-    
-    init(from decoder: Decoder) throws {
-        if var container = try? decoder.unkeyedContainer() {
-            var value: [String] = []
-            while !container.isAtEnd {
-                if let v = try? container.decode(String.self) {
-                    value.append(v)
-                }
-            }
-            self.value = .list(value)
-        } else {
-            let container = try decoder.singleValueContainer()
-            self.value = .string(try container.decode(String.self))
-        }
-    }
-    
-    init(stringLiteral: String) {
-        self.value = .string(stringLiteral)
-    }
-    
-    init(arrayLiteral elements: String...) {
-        self.value = .list(elements)
-    }
-    
-    func stringValue() -> String {
-        if case let .string(s) = value {
-            return s
-        }
-        
-        return ""
-    }
-    
-    func listValue() -> [String] {
-        if case let .list(l) = value {
-            return l
-        } else if case let .string(s) = value {
-            return [s]
-        }
-        
-        return []
-    }
-}
-
-typealias SettingList = [String]
-typealias SettingsDictionary = [String:SettingsValue]
-
-// TODO: this should really be read from a json file or series of json files (one per tool/platform?)
-// TODO: it would also make sense to allow users to add/contribute to it somehow
-let settingsMapping : [String:Any] = [
-    "minimum-target": [
-        "swift" : [
-            "prefix": ["-Xswiftc", "-target", "-Xswiftc"],
-            "values": [
-                "macosx10.12" : "x86_64-apple-macosx10.12",
-                "macosx10.13" : "x86_64-apple-macosx10.13"
-            ]
-        ],
-        "xcconfig": [
-            "prefix": ["MACOSX_DEPLOYMENT_TARGET = "],
-            "values" : [
-                "macosx10.12": "10.12",
-                "macosx10.13": "10.13"
-            ]
-        ]
-    ],
-    "optimisation": [
-        "swift" : [
-            "prefix": ["-Xswiftc"],
-            "values": [
-                "none": "-Onone",
-                "size": "-Osize",
-                "speed": "-O"
-            ]
-        ],
-        "xcconfig" : [
-            "prefix" : ["SWIFT_OPTIMIZATION_LEVEL = "],
-            "values": [
-                "none": "-Onone",
-                "size": "-Osize",
-                "speed": "-O"
-            ]
-        ]
-    ],
-    
-    "definition": [
-        "swift" : [
-            "rawValues" : ["-Xswiftc", "-D"]
-        ],
-    ]
-]
 
 
-struct Inheritance : Decodable {
-    let name : String
-    let filter : [String]?
-}
 
-struct Settings : Decodable {
-    let values : SettingsDictionary?
-    let inherits : [Inheritance]?
-    
-    func mappedSettings(for name: String, value: SettingsValue, mapping: [String:Any]) -> [String] {
-        var result: [String] = []
-        if let prefix = mapping["prefix"] as? [String] {
-            result.append(contentsOf: prefix)
-        }
-        
-        if let valueMappings = mapping["values"] as? [String:Any] {
-            for value in value.listValue() {
-                if let value = valueMappings[value] as? String {
-                    result.append(value)
-                } else if let values = valueMappings[value] as? [String] {
-                    result.append(contentsOf: values)
-                }
-            }
-        } else if let rawValues = mapping["rawValues"] as? [String] {
-            var prefixValues = rawValues
-            prefixValues.removeLast()
-            if let lastValue = rawValues.last {
-                for value in value.listValue() {
-                    result.append(contentsOf: prefixValues)
-                    result.append(lastValue.appending(value))
-                }
-            }
-        }
-
-        return result
-    }
-    
-    func mappedSettings(for tool: String) -> [String] {
-        var args : [String] = []
-        
-        if let values = values {
-            for value in values {
-                if let mapping = settingsMapping[value.key] as? [String:Any] {
-                    if let toolMapping = mapping[tool] as? [String:Any] {
-                        args.append(contentsOf: mappedSettings(for: value.key, value: value.value, mapping: toolMapping))
-                    }
-                }
-            }
-        }
-
-        return args
-    }
-    
-    static func mergedLists(_ l1 : SettingList?, _ l2 : SettingList?) -> SettingList {
-        if l1 == nil {
-            if l2 == nil {
-                return []
-            } else {
-                return l2!
-            }
-        } else {
-            if l2 == nil {
-                return l1!
-            } else {
-                return l1! + l2!
-            }
-        }
-    }
-
-    static func mergedDictionaries(_ l1 : SettingsDictionary?, _ l2 : SettingsDictionary?) -> SettingsDictionary {
-        if l1 == nil {
-            if l2 == nil {
-                return [:]
-            } else {
-                return l2!
-            }
-        } else {
-            if l2 == nil {
-                return l1!
-            } else {
-                return l1!.merging(l2!, uniquingKeysWith: { (l,r) in return l })
-            }
-        }
-    }
-
-    static func mergedSettings(_ s1 : Settings, _ s2 : Settings) -> Settings {
-        return Settings(
-            values: mergedDictionaries(s1.values, s2.values),
-            inherits: nil)
-    }
-}
 
 /**
  Data structure representing a phase of the build process.
